consolidate compact QMDB code
