+++ storage/src/qmdb/any/batch.rs
@@ -14,11 +14,10 @@ use crate::{
             ordered::{find_next_key, find_prev_key},
             ValueEncoding,
         },
-        append_batch::BatchBounds,
+        batch::{BatchBounds, Plan},
         bitmap::Shared,
         delete_known_loc,
         operation::{Key, Operation as OperationTrait},
-        plan::Plan,
         update_known_loc,
     },
     Context,
