+++ storage/src/qmdb/compact/mod.rs
@@ -1,0 +1,10 @@
+pub(crate) mod db;
+pub(crate) mod witness;
+
+pub(crate) use db::CompactDbInner;
+#[cfg(test)]
+pub(crate) use witness::{commit_op_key, proof_key};
+pub(crate) use witness::{
+    init_compact_witness, load_active_witness, persist_cached_witness, persist_witness,
+    validate_witness, CompactCommit, Witness, WitnessSource,
+};
