+++ storage/src/qmdb/keyless/compact.rs
@@ -25,10 +25,8 @@ use crate::{
     },
     qmdb::{
         any::value::ValueEncoding,
-        append_batch::{AppendBatchView, BatchBounds, CompactBatch},
-        compact_db::CompactDbInner,
-        compact_witness::CompactCommit,
-        plan::apply,
+        batch::{apply, AppendBatchView, BatchBounds, CompactBatch},
+        compact::{CompactCommit, CompactDbInner},
         sync::compact as compact_sync,
         Error,
     },
@@ -766,7 +764,7 @@ mod tests {
             tamper_metadata_key(
                 context.with_label("tamper"),
                 partition,
-                crate::qmdb::compact_witness::proof_key(slot),
+                crate::qmdb::compact::proof_key(slot),
             )
             .await;
 
@@ -805,7 +803,7 @@ mod tests {
             overwrite_metadata_key(
                 context.with_label("tamper"),
                 partition,
-                crate::qmdb::compact_witness::commit_op_key(slot),
+                crate::qmdb::compact::commit_op_key(slot),
                 Operation::<mmr::Family, FixedEncoding<U64>>::Commit(
                     Some(U64::new(11)),
                     oversized_floor,
